flash minimized and on top
