+++ components/atoms/FlashInfo.module.css
@@ -13,7 +13,7 @@
     align-items: center;
     justify-content: space-between;
     margin: 0 auto;
-    padding: 16px 0;
+    padding: 8px 0;
 }
 
 .flash_inner__link {
