+++ components/atoms/Layout.module.css
@@ -1,3 +1,7 @@
 .container {
     position: relative;
 }
+
+.container_flash {
+    margin-top: 52px;
+}
