+++ pages/index.js
@@ -9,24 +9,12 @@ import { useRouter } from 'next/router'
 import PropTypes from 'prop-types'
 import React, { useEffect, useState } from 'react'
 
-const infos = [
-  {
-    label: 'dim 08 nov 2020 / VTT, FTT et Handbike / Handisport',
-    redirect: '/',
-  },
-  {
-    label: 'dim 15 nov 2020 / Sortie VTT – 25 à 35km / VTT',
-    redirect: '/',
-  },
-]
-
 Home.propTypes = {
   stories: PropTypes.array,
   partners: PropTypes.array,
 }
 
 function Home({ stories, partners }) {
-  const [flash, setFlash] = useState(true)
   const router = useRouter()
 
   const handleStoriesClub = () => {
@@ -37,54 +25,18 @@ function Home({ stories, partners }) {
     router.push('/stories')
   }
 
-  useEffect(() => {
-    const item = window.localStorage.getItem('flash')
-
-    if (!item) {
-      setFlash(true)
-      window.localStorage.setItem(
-        'flash',
-        JSON.stringify({
-          value: true,
-          createdAt: new Date(),
-        }),
-      )
-    } else {
-      const { value, createdAt } = JSON.parse(item)
-      const oneDay = 86400
-      const isExpired =
-        new Date(createdAt).getTime() - new Date().getTime() > oneDay
-      if (isExpired) {
-        window.localStorage.removeItem('flash')
-      } else if (!value) {
-        setFlash(false)
-      }
-    }
-  }, [])
-
-  const toggleFlash = () => {
-    setFlash(false)
-    window.localStorage.removeItem('flash')
-    window.localStorage.setItem(
-      'flash',
-      JSON.stringify({
-        value: false,
-        createdAt: new Date(),
-      }),
-    )
-  }
-
   return (
-    <Layout>
-      {flash && <FlashInfo infos={infos} handleClose={() => toggleFlash()} />}
-      <LandingHero handleRedirection={handleStoriesClub} />
-      <LandingStoriesHighlight
-        highlightedStories={stories}
-        handleRedirection={handleStoriesRedirection}
-      />
+    <>
+      <Layout>
+        <LandingHero handleRedirection={handleStoriesClub} />
+        <LandingStoriesHighlight
+          highlightedStories={stories}
+          handleRedirection={handleStoriesRedirection}
+        />
 
-      <LandingContact partners={partners} />
-    </Layout>
+        <LandingContact partners={partners} />
+      </Layout>
+    </>
   )
 }
